Moved functions into state that belong in state
